feat(dev-tools): extend toolbar for preview envs and add resource warnings tab
- Enable DevToolbar in Vercel preview deployments by checking
  NEXT_PUBLIC_VERCEL_ENV === 'preview' alongside NODE_ENV === 'development'
  in all guards and the index.ts tree-shaking condition
- Rename "Dev Telemetry" → "Dev Toolbar"; badge shows "Preview" or "Local"
- Add ExtraTab extension point to DevToolbar so consumers can inject tabs
- Add ResourceWarningsTab in studio: same off/warn/crit toggles for all
  resource warning types, wired into the existing DevToolbar via _app.tsx
- Remove standalone SupabaseDevToolbar (superseded by the tab above)
- localStorage persistence already handled by the existing STORAGE_KEY
  logic in DevToolbarContext — enabled state survives refreshes
